LayoutTests/imported/w3c/web-platform-tests/reporting/cross-origin-report-no-credentials.https.sub-expected.txt

CONSOLE MESSAGE: Refused to load https://web-platform.test:9443/reporting/resources/fail.png because it does not appear in the img-src directive of the Content Security Policy.


PASS Reporting endpoints did not receive credentials.
